synchronisation upstream 1.52

+++ www/export_csv.php
@@ -205,7 +205,7 @@ while($lineTmp = $lines->fetch()) {
 		}
 
 		$dateDebut_projet = new DateTime();
-		$dateDebut_projet->setDate(substr($periode->date_debut,0,4), substr($periode->date_debut,5,2), substr($periode->date_debut,8,2));
+		$dateDebut_projet->setDate((int)substr($periode->date_debut,0,4), (int)substr($periode->date_debut,5,2), (int)substr($periode->date_debut,8,2));
 
 		$dateFin_projet = new DateTime();
 
@@ -214,7 +214,7 @@ while($lineTmp = $lines->fetch()) {
 			$dateFin_projet = clone $dateDebut_projet;
 		}
 		else {
-			$dateFin_projet->setDate(substr($periode->date_fin,0,4), substr($periode->date_fin,5,2), substr($periode->date_fin,8,2));
+			$dateFin_projet->setDate((int)substr($periode->date_fin,0,4), (int)substr($periode->date_fin,5,2), (int)substr($periode->date_fin,8,2));
 		}
 
 		while ($tmpDate <= $dateFin_projet) {
