+++ www/export_csv_raw.php
@@ -74,10 +74,10 @@ $periodes->db_loadSQL($sql);
 while ($periode = $periodes->fetch()) {
 	if (!is_null($periode->date_fin)) {
 		$dateDebut_projet = new DateTime();
-		$dateDebut_projet->setDate(substr($periode->date_debut,0,4), substr($periode->date_debut,5,2), substr($periode->date_debut,8,2));
+		$dateDebut_projet->setDate((int)substr($periode->date_debut,0,4), (int)substr($periode->date_debut,5,2), (int)substr($periode->date_debut,8,2));
 		$dateFin_projet = new DateTime();
 		$tmpDate = clone $dateDebut_projet;
-		$dateFin_projet->setDate(substr($periode->date_fin,0,4), substr($periode->date_fin,5,2), substr($periode->date_fin,8,2));
+		$dateFin_projet->setDate((int)substr($periode->date_fin,0,4), (int)substr($periode->date_fin,5,2), (int)substr($periode->date_fin,8,2));
 
 		
 		while ($tmpDate <= $dateFin_projet) {
