+++ www/export_pdf_calendrier.php
@@ -221,7 +221,7 @@ while($lineTmp = $lines->fetch()) {
 		}
 
 		$dateDebut_projet = new DateTime();
-		$dateDebut_projet->setDate(substr($periode->date_debut,0,4), substr($periode->date_debut,5,2), substr($periode->date_debut,8,2));
+		$dateDebut_projet->setDate((int)substr($periode->date_debut,0,4), (int)substr($periode->date_debut,5,2), (int)substr($periode->date_debut,8,2));
 
 		$dateFin_projet = new DateTime();
 
@@ -230,7 +230,7 @@ while($lineTmp = $lines->fetch()) {
 			$dateFin_projet = clone $dateDebut_projet;
 		}
 		else {
-			$dateFin_projet->setDate(substr($periode->date_fin,0,4), substr($periode->date_fin,5,2), substr($periode->date_fin,8,2));
+			$dateFin_projet->setDate((int)substr($periode->date_fin,0,4), (int)substr($periode->date_fin,5,2), (int)substr($periode->date_fin,8,2));
 		}
 
 		while ($tmpDate <= $dateFin_projet) {
